treminal commands appended

+++ Transaction.sol
@@ -3,22 +3,22 @@ pragma solidity ^0.4.11;
 
 contract Transaction {
   
-  mapping (bytes32 => uint8) public candidateCredit;
-  mapping (bytes32 => uint8) public candidateDebt;
+  mapping (bytes32 => uint32) public candidateCredit;
+  mapping (bytes32 => uint32) public candidateDebt;
   
   bytes32[] public candidateList;
   function Transaction(bytes32[] candidateNames) {
     candidateList = candidateNames;
   }
 
-  function getCredit(bytes32 candidate) returns (uint8) {
+  function getCredit(bytes32 candidate) returns (uint32) {
     return candidateCredit[candidate];
   }
 
-  function getDebt(bytes32 candidate) returns (uint8) {
+  function getDebt(bytes32 candidate) returns (uint32) {
       return candidateDebt[candidate];
   }
-  function sendMoney(bytes32 candidate, uint8 money) {
+  function sendMoney(bytes32 candidate, uint32 money) {
     if (candidateDebt[candidate] > money) {
       candidateDebt[candidate] -= money;
     } else {
